+++ src/com/intellij/idea/plugin/hybris/system/type/codeInsight/daemon/ModelEnumLineMarkerProvider.kt
@@ -22,7 +22,7 @@ import com.intellij.idea.plugin.hybris.codeInsight.daemon.AbstractHybrisClassLin
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.system.type.meta.TSMetaModelAccess
-import com.intellij.idea.plugin.hybris.system.type.util.ModelsUtils
+import com.intellij.idea.plugin.hybris.system.type.util.TSUtils
 import com.intellij.openapi.editor.markup.GutterIconRenderer
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
@@ -32,7 +32,7 @@ class ModelEnumLineMarkerProvider : AbstractHybrisClassLineMarkerProvider<PsiCla
 
     override fun getName() = message("hybris.editor.gutter.ts.model.enum.name")
     override fun getIcon(): Icon = HybrisIcons.TS_ENUM
-    override fun canProcess(psi: PsiClass) = ModelsUtils.isEnumFile(psi)
+    override fun canProcess(psi: PsiClass) = TSUtils.isEnumFile(psi)
     override fun tryCast(psi: PsiElement) = (psi as? PsiClass)
         ?.takeIf { it.nameIdentifier != null }