+++ src/com/intellij/idea/plugin/hybris/toolwindow/ccv2/views/CCv2DeploymentsDataView.kt
@@ -67,7 +67,8 @@ object CCv2DeploymentsDataView : AbstractCCv2DataView<CCv2Deployment>() {
 
             panel {
                 row {
-                    label(deployment.status)
+                    icon(deployment.status.icon)
+                    label(deployment.status.title)
                         .comment("Status")
                 }
             }.gap(RightGap.COLUMNS)